requirements.txt
# ==============================================================================
# Core Framework
# ==============================================================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ==============================================================================
# AI / LLM - LangChain Orchestra
# ==============================================================================
langchain==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-community==0.3.9
langgraph==0.2.53
langchain-text-splitters==0.3.2

# ==============================================================================
# Azure AI Inference (Phi, Mistral, Llama via Azure)
# ==============================================================================
azure-ai-inference==1.0.0b7
azure-identity==1.19.0
azure-core==1.32.0

# ==============================================================================
# OpenAI SDK directo
# ==============================================================================
openai==1.57.0

# ==============================================================================
# RAG - Vector Stores & Embeddings
# ==============================================================================
langchain-postgres>=0.0.17
pgvector>=0.2.5
faiss-cpu>=1.9.0

# ==============================================================================
# Database - PostgreSQL Async
# ==============================================================================
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg[binary,pool]==3.2.4

# ==============================================================================
# HTTP & Streaming
# ==============================================================================
httpx==0.28.1
sse-starlette==2.1.3

# ==============================================================================
# Observability & Logging
# ==============================================================================
structlog==24.4.0
python-json-logger==3.2.0

# ==============================================================================
# Utilities
# ==============================================================================
tenacity==9.0.0
cachetools==5.5.0
python-multipart==0.0.20
aiofiles==24.1.0

# ==============================================================================
# Dev / Testing
# ==============================================================================
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
black==24.10.0
ruff==0.8.4
